+++ wiki-app/search_hash.h
@@ -17,6 +17,9 @@
 
 #ifndef SEARCH_HASH_H
 #define SEARCH_HASH_H
+
+#include <inttypes.h>
+
 #include "bigram.h"
 
 #define MAX_SEARCH_HASH_TABLE_ENTRIES (4 * 256 * 1024)
@@ -26,8 +29,8 @@
 #define SEARCH_HASH_SEQUENTIAL_SEARCH_THRESHOLD 64
 
 typedef struct _search_hash_table {
-	long offset_fnd; // offset to pedia.fnd for the search title hashed
-	long next_entry_idx; // byte 1: length of hash string byte 2~4: index (of struct hash_table array) of the next entry with the same hash key
+	uint32_t offset_fnd; // offset to pedia.fnd for the search title hashed
+	uint32_t next_entry_idx; // byte 1: length of hash string byte 2~4: index (of struct hash_table array) of the next entry with the same hash key
 } SEARCH_HASH_TABLE;
 
 typedef struct _search_hash_string {
